Refactoring finished.
